Remove various io/os functions considered dangerous to overall application state.
Force stdin/stdout to be "closed" and unavailable.

tag any use of filepaths with a check for validity in the sandbox (eg, lets not let people delete Win32)
forcing that the filepaths provided are relative to the scripts folder and forced to be subpaths within
that directory, resolving soft/hard links appropriately, and ensuring the containing directory exists.
handles .. in paths as well.

paths (relative to io directory which can be "anything") are subject to constraints that make the paths "simple"
and work on windows and linux alike. paths are also forced to be lower-case to avoid issues between linux/windows
where windows is not case-sensitive on file accesses.